Move program enrollment views to enrollment_views.py
