+++ lms/djangoapps/program_enrollments/api/v1/urls.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 from django.conf.urls import url
 
 from lms.djangoapps.program_enrollments.api.v1.constants import PROGRAM_UUID_PATTERN
-from lms.djangoapps.program_enrollments.api.v1.views import (
+from lms.djangoapps.program_enrollments.api.v1.enrollment_views import (
     ProgramEnrollmentsView,
     ProgramCourseEnrollmentsView,
     ProgramCourseEnrollmentOverviewView,
